시큐리티 추가

+++ src/main/java/kr/kwangan2/springbootboard/board/controller/BoardController.java
@@ -6,11 +6,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import kr.kwangan2.springbootboard.board.entity.Board;
 import kr.kwangan2.springbootboard.board.service.BoardService;
+import kr.kwangan2.springbootboard.member.entity.Member;
 
 @Controller
 @RequestMapping("/board")
@@ -19,50 +21,72 @@ public class BoardController{
 	@Autowired
 	private BoardService boardService;
 	
-	@GetMapping("/listBoard")
-	public String listBoard(Model model, Board board) {
+	@ModelAttribute("member")
+	public Member setMemer() {
+		return new Member();
+	}
+	
+	@RequestMapping("/listBoard")
+	public String listBoard(Model model, @ModelAttribute("member") Member member, Board board) {
+		if(member.getId() == null) {
+			return "redirect:/member/login";
+		}
 		List<Board> boardList = boardService.listBoard(board);
-//		for (int i=1; i<=10; i++) {
-//			Board board = new Board("제목 ["+i+"]", "작성자"+i, "게시판내용["+i+"]", new Date(), 0L);
-//			boardList.add(board);
-//		}
 		model.addAttribute("boardList", boardList);
 		return "/board/listBoard";
 	}
 
     @GetMapping("/getBoard")
-    public String getBoard(Board board, Model model, String action) {
+    public String getBoard(@ModelAttribute("member") Member member, Board board, Model model, String action) {
+		if(member.getId() == null) {
+			return "redirect:/member/login";
+		}
          model.addAttribute("board", boardService.getBoard(board,action));
       return "/board/getBoard";
     }
 
 	
 	@GetMapping("/insertBoard")
-	public String insertBoard(Board board) {
+	public String insertBoard(@ModelAttribute("member") Member member, Board board) {
+		if(member.getId() == null) {
+			return "redirect:/member/login";
+		}
 		return "/board/insertBoard";
 	}
 	
 	
 	@PostMapping("/insertBoardProc")
-	public String insertBoardProc(Board board) {
+	public String insertBoardProc(@ModelAttribute("member") Member member, Board board) {
+		if(member.getId() == null) {
+			return "redirect:/member/login";
+		}
 		boardService.insertBoard(board);
 		return "redirect:/board/listBoard";
 	}
 	
 	@PostMapping("/updateBoardProc")
-	public String updateBoard(Board board) {
+	public String updateBoard(@ModelAttribute("member") Member member, Board board) {
+		if(member.getId() == null) {
+			return "redirect:/member/login";
+		}
 		boardService.updateBoard(board);
 		return "redirect:/board/listBoard";
 	}
 	
 	@GetMapping("/deleteBoard")
-	public String deleteBoard(Board board) {
+	public String deleteBoard(@ModelAttribute("member") Member member, Board board) {
+		if(member.getId() == null) {
+			return "redirect:/member/login";
+		}
 		boardService.deleteBoard(board);
 		return "redirect:/board/listBoard";
 	}
 	
 	@GetMapping("/hello")
-	public String hello(Model model){
+	public String hello(@ModelAttribute("member") Member member, Model model){
+		if(member.getId() == null) {
+			return "redirect:/member/login";
+		}
 	model.addAttribute("greeting", "Hello Thymleaf!!!");
 	return "/board/hello";
 	}
